web-scraper/requirements.txt
amqp==2.5.2
asgiref==3.2.5
beautifulsoup4==4.8.2
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-jsonfield==1.4.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
lxml==4.5.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
responses==0.10.12
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
